Update Vue and MomentJS CDN versions and add versioning in composer.json
The Vue and MomentJS CDN import links were updated in the logs.latte file to point to the major version (2) without specifying the minor versions. This would allow for automatic updates within the same major version. Meanwhile in composer.json, versioning was added to properly track the current version of the module.