Refactor log deletion
Refactoring was performed on the handleDelete function in LogsControl.php to improve the deletion of log files. Previously, the function was using shell commands to delete log files which was not performant and prone to errors. Now, PHP's unlink function is used alongside glob to identify and remove log files, improving performance and maintainability.